fix(fg-stock): keep cage qty decimal in location snapshot (restores Σcards==Σphysique)
fg_stock_location_snapshot() floored the cage anchor qty with (int), while
fg_stock_compute() kept it decimal — so the 7 decimal cage-unit anchors
(ZEPX 0.2171 … STIX 1.0078, mig 363) showed 0 on per-location cards and
Σcards fell short of Σphysique by 2.4655. (int)→(float). Verified: per-SKU
divergence 0, Σcards==Σphysique==7651.4655 at flag-OFF and flag-ON.

Prereq for activating the repack-depletion leg (repack_depletion_live) whose
R1 invariant is exactly Σcards==Σphysique.

+++ app/fg-stock.php
@@ -1362,7 +1362,7 @@ function fg_stock_location_snapshot(PDO $pdo): array
                 'format'         => $ar['format'],
                 'display_family' => $ar['display_family'],
                 'hl_per_unit'    => (float) $ar['hl_per_unit'],
-                'qty'            => (int) $ar['qty'],
+                'qty'            => (float) $ar['qty'],   // float, NOT (int): cages are decimal cage-units (mig 363); flooring here broke Σcards==Σphysique
                 'counted_at'     => $ar['counted_at'],
             ];
         }
